% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_Explore_1536.R
\docType{data}
\name{Olink_Explore_1536}
\alias{Olink_Explore_1536}
\title{Olink/Explore 1536 panel}
\format{
A data frame with 1,472 rows and 3 variables:
\describe{
\item{\code{UniProt}}{UniProt id}
\item{\code{Assay}}{Experimental assay}
\item{\code{Panel}}{Olink panel}
}
}
\usage{
data(Olink_Explore_1536)
}
\description{
Information based on pilot studies
}
\details{
Curated from R.
}
\keyword{datasets}
\keyword{internal}
