% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_qPCR.R
\docType{data}
\name{Olink_qPCR}
\alias{Olink_qPCR}
\title{Olink/qPCR panels}
\format{
A data frame with 1,112 rows and 7 variables:
\describe{
\item{\code{UniProt}}{UniProt id}
\item{\code{Panel}}{Panels}
\item{\code{Target}}{Protein}
\item{\code{gene}}{HGNC symbol}
\item{\code{chr}}{Chromosome}
\item{\code{start}}{start}
\item{\code{end}}{end}
}
}
\usage{
data(Olink_qPCR)
}
\description{
Information on all qPCR panels
}
\details{
Curated from Excel.
}
\keyword{datasets}
\keyword{internal}
