% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scallop_inf1.R
\docType{data}
\name{scallop_inf1}
\alias{scallop_inf1}
\title{Supplementary table 3}
\format{
A data frame with 180 rows and 19 variables:
\describe{
\item{UniProt}{UnitProt ID}
\item{Protein}{Protein name}
\item{Protein_gene_symbol}{Gene symbol}
\item{Chromosome}{Chromosome}
\item{Position}{Position}
\item{cistrans}{cis/trans}
\item{rsid}{reference sequence ID}
\item{Effect_allele}{Effect allele}
\item{Other_allele}{Eeference allele}
\item{EAF}{Effect allele frequency}
\item{b}{b}
\item{SE}{SE}
\item{log10P}{log10(P)}
\item{Direction}{Direction field in METAL output}
\item{HetISq}{\emph{I}\eqn{^2}}
\item{HetChiSq}{Heterogeneity chi-square}
\item{HetDf}{degrees of freedom}
\item{logHetP}{Heterogeneity log10(P)}
\item{N}{N}
}
}
\usage{
data(scallop_inf1)
}
\description{
Supplementary information for \insertCite{zhao23;textual}{pQTLdata}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
\keyword{internal}
