% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer1980.R
\docType{data}
\name{seer1980}
\alias{seer1980}
\title{Seer 1980 panel}
\format{
A data frame with 1,980 rows:
\describe{
\item{\code{PID.NP}}{PID.NP}
\item{\code{protein_ids}}{protein_ids}
\item{\code{protein_names}}{protein_names}
\item{\code{mapped.UniProtID}}{mapped.UniProtID}
\item{\code{mapped_gene_id}}{mapped_gene_id}
\item{\code{gene_name}}{gene_name}
\item{\code{description}}{description}
\item{\code{chr}}{chr}
\item{\code{start}}{start}
\item{\code{end}}{end}
}
}
\usage{
data(seer1980)
}
\description{
ST1 from \insertCite{suhre24;textual}{pQTLdata}.
}
\details{
As above.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
\keyword{internal}
