% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SomaScanV4.1.R
\docType{data}
\name{SomaScanV4.1}
\alias{SomaScanV4.1}
\title{SomaScan v4.1}
\format{
A data frame with 7,288 rows and 6 variables:
\describe{
\item{\code{#}}{A serial number}
\item{\code{SeqID}}{SeqID}
\item{\code{Human.Target.or.Analyte}}{Human target/analyte}
\item{\code{UniProt.ID}}{UniProt id}
\item{\code{GeneID}}{HGNC symbol}
\item{\code{Type}}{"Protein"}
}
}
\usage{
data(SomaScanV4.1)
}
\description{
This is the 7k panel
}
\details{
obtained directly from SomaLogic.
}
\keyword{datasets}
\keyword{internal}
