% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_Explore_3072.R
\docType{data}
\name{Olink_Explore_3072}
\alias{Olink_Explore_3072}
\title{Olink/Explore 3072 panels}
\format{
A data frame with 2945 rows and 4 variables:
\describe{
\item{\code{UniProt.ID}}{UniProt id}
\item{\code{Protein.name}}{Protein name}
\item{\code{Gene.name}}{Gene name}
\item{\code{Explore.384.panel}}{Explore 384 panel}
}
}
\usage{
Olink_Explore_3072
}
\description{
Information on all qPCR panels
}
\details{
Curated from Excel.
}
\keyword{datasets}
\keyword{internal}
