% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{Olink_Explore_HT}
\alias{Olink_Explore_HT}
\title{Olink/Explore HT panels}
\format{
A data frame with 5343 rows and 3 variables:
\describe{
\item{\code{UniProt.ID}}{UniProt id}
\item{\code{Protein.name}}{Protein name}
\item{\code{Gene.name}}{Gene name}
}
}
\usage{
Olink_Explore_HT
}
\description{
Information on all qPCR panels
}
\details{
Curated from Excel.
}
\keyword{datasets}
\keyword{internal}
