% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_Target_96.R
\docType{data}
\name{Olink_Target_96}
\alias{Olink_Target_96}
\title{Olink/Target 96 panels}
\format{
A data frame with 1,116 rows and 3 variables:
\describe{
\item{\code{UniProt}}{UniProt id}
\item{\code{Target}}{Protein}
\item{\code{Panel}}{Panel}
}
}
\usage{
Olink_Target_96
}
\description{
Information on all Target 96 panels. Individual panels are also available from the companion xlsx in the Olink/ directory.
}
\details{
Curated from Excel.
}
\keyword{datasets}
\keyword{internal}
