% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{SomaScan11k}
\alias{SomaScan11k}
\title{SomaScan 11k}
\format{
A data frame with 7288 rows and 6 variables:
\describe{
\item{\code{Sequence.ID}}{Sequence ID}
\item{\code{Full.Name}}{Full name}
\item{\code{Target.Name}}{Target name}
\item{\code{UniProt.ID}}{UniProt ID}
\item{\code{Entrez.Gene.Name}}{Entrez gene name}
}
}
\source{
https://somalogic.com/somascan-11k-assay/
}
\usage{
SomaScan11k
}
\description{
This is also the latest panel
}
\details{
obtained directly from SomaLogic.
}
\keyword{datasets}
\keyword{internal}
