% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swath_ms.R
\docType{data}
\name{swath_ms}
\alias{swath_ms}
\title{SWATH-MS panel}
\format{
A data frame with 684 rows and 5 variables:
\describe{
\item{\code{Accession}}{UniProt id}
\item{\code{accList}}{List of UniProt ids}
\item{\code{uniprotName}}{Protein}
\item{\code{ensGene}}{ENSEMBL gene}
\item{\code{geneName}}{HGNC symbol}
}
}
\usage{
swath_ms
}
\description{
As indicated by \insertCite{ludwig18;textual}{pQTLdata}, Sequential Windowed Acquisition
of All Theoretical Fragment Ion Mass Spectra (SWATH-MS) represents a specific variant of
Data-Independent Acquisition (DIA).
The panel shown here is curated during INTERVAL pilot study.
}
\details{
As above.
}
\keyword{datasets}
\keyword{internal}
